@@ -14,10 +14,9 @@
     <name>native-spring-api</name>
     <description>Demo project for Spring Boot</description>
     <properties>
-        <java.version>11</java.version>
-        <kotlin.version>1.4.30</kotlin.version>
-        <spring-native.version>0.9.0</spring-native.version>
-        <testcontainers.version>1.15.2</testcontainers.version>
+        <java.version>17</java.version>
+        <repackage.classifier/>
+        <testcontainers.version>1.17.6</testcontainers.version>
     </properties>
     <dependencies>
         <dependency>
