@@ -6,15 +6,15 @@ import com.fasterxml.jackson.annotation.ObjectIdGenerators
 import org.hibernate.annotations.GenericGenerator
 import org.springframework.data.annotation.Id
 import java.util.*
-import javax.persistence.Entity
-import javax.persistence.GeneratedValue
-import javax.persistence.ManyToMany
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.ManyToMany
 
 @Entity
 @JsonIdentityInfo(generator= ObjectIdGenerators.PropertyGenerator::class, property="id")
 class Movie {
 
-    @javax.persistence.Id
+    @jakarta.persistence.Id
     @Id
     @GeneratedValue(generator = "uuid")
     @GenericGenerator(name = "uuid", strategy = "uuid2")
