@@ -2,10 +2,10 @@ package com.lafabrique.digit.owl
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
-import java.util.*
+import java.util.UUID
 
 @Repository
 interface MovieRepository : JpaRepository<Movie, UUID> {
 
-    fun findByTitle(title: String): List<Movie>
+//    fun findByTitle(title: String): List<Movie>
 }
